Move paging from hal to library
